Drop middleware sub-package
